fix: Fixed Piano.
Before this commit the piano keys where wrong assigned. Now the whole piano works fine and is optimised. Before, the way we had to identify which keyboard key was pressed was the keyChar, but as there were some keys that didn't had a representative char in this commit it has been changed so the keyCode is the identifier used.