feat: Added SongList and Friend Views and Controllers.
Fixed the Friends button on MainMenuView and now leads to FriendView.

Added FriendView FriendController the buttons on the view are not implemented yet just the SongList button and the back to menu button.

Added SongListView SongListController the buttons on the view are not implemented yet just the PlaySong button and the back to FriendView button.