fix: Client loss error.
Before this commit, the program started after a login, which created a client class that connected to the server, which allowed the menu manager to know that Client, so if the user wanted to make a friend request, the menu manager would send the client to the friendList manager and everything would've worked. But if before doing this you entered the piano, the MainMenuManager did not send the client to the piano, so when you'd leave the piano you'd not have the client in the MainMenuManager. This caused that if you wanted to make a request to the server it would give nullPointerException, since the client had been lost. Now all managers have the client.