Feat: Save song to the server.
A new view to save the song has been created. Now, the song name, album and privacy is stored in the server. However, the songfile.mid is stored locally in the root project. Not in the server.

Fixed: Client lost through views. Before this commit, the program crashed when going to the main menu from the friend list song. Now is fixed.